+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.42.3](https://github.com/entrecode/ec.components/compare/@ec.components/ui@0.42.2...@ec.components/ui@0.42.3) (2019-04-12)
+
+
+### Bug Fixes
+
+* EDITOR-330 => ids are now filterable ([ee3518f](https://github.com/entrecode/ec.components/commit/ee3518f))
+
+
+
+
+
 ## [0.42.2](https://github.com/entrecode/ec.components/compare/@ec.components/ui@0.42.1...@ec.components/ui@0.42.2) (2019-04-08)
 
 **Note:** Version bump only for package @ec.components/ui